chore: use theme colors
