+++ apps/catalogue/app/components/store/HeaderButton.vue
@@ -1,7 +1,7 @@
 <template>
   <Button label="Request " icon="shopping-cart" @click="$emit('click')">
     <span class="inline-block w-6 h-6 bg-white rounded-full">
-      <span class="body-xs/1">
+      <span class="body-xs/1 text-black">
         {{ numberOfItemsInStore }}
       </span>
     </span>
